# api/docs/swagger.yaml
definitions:
  structures.NoteStruct:
    properties:
      body:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      exp_time:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  structures.UserStruct:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/create_user/:
    post:
      consumes:
      - application/json
      description: This API for creating user
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structures.UserStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.UserStruct'
      summary: Create user
      tags:
      - Users
  /v1/createnote/:
    post:
      consumes:
      - application/json
      description: This api for Creating Note
      parameters:
      - description: note body
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/structures.NoteStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create Note
      tags:
      - Note
  /v1/delete_user/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting user
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete user
      tags:
      - Users
  /v1/deletenote/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting note
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete Note
      tags:
      - Note
  /v1/notes/:
    post:
      consumes:
      - application/json
      description: This api for Seting Notes
      parameters:
      - description: exptime
        in: query
        name: exptime
        required: true
        type: integer
      - description: note body
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/structures.NoteStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Set Note
      tags:
      - Note
  /v1/update_user/:
    post:
      consumes:
      - application/json
      description: This API for updating user
      parameters:
      - description: update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structures.UserStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update user
      tags:
      - Users
  /v1/updatenote/:
    put:
      consumes:
      - application/json
      description: This api for Updating Note
      parameters:
      - description: note body
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/structures.NoteStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update Note
      tags:
      - Note
swagger: "2.0"